online_training/serializers/subscription.py
from rest_framework import serializers

from online_training.models import Subscription


class SubscriptionSerializer(serializers.ModelSerializer):

    class Meta:
        model = Subscription
        fields = '__all__'