pyproject.toml
[tool.poetry]
name = "drf-online-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "drf_online_training"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
ipython = "^8.20.0"
django-dump-load-utf8 = "^0.0.4"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
routers = "^0.10.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
generics = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
